Extract standard test systems into test_utils
